:truck: Move input classes to own directory
